Extract shared core package
