+++ app/__init__.py
@@ -4,10 +4,9 @@ from flask import Flask
 from flask_cors import CORS
 
 from app.api import register_blueprints
-from app.config import Config
-from app.extensions import db, cache, migrate, scheduler
-from app.logging_config import configure_logging
-from app.tasks.sync import register_jobs
+from coclib.config import Config
+from coclib.extensions import db, cache, migrate, scheduler
+from coclib.logging_config import configure_logging
 
 logger = logging.getLogger(__name__)
 
@@ -22,12 +21,10 @@ def create_app(cfg_cls: type[Config] = Config) -> Flask:
     db.init_app(app)
     cache.init_app(app)
     scheduler.init_app(app)
-    scheduler.start()
     migrate.init_app(app, db)
 
     register_blueprints(app)
 
-    with app.app_context():
-        register_jobs()
+
 
     return app
