+++ app/api/player_routes.py
@@ -1,6 +1,6 @@
 from flask import Blueprint, jsonify
-from app.services.player_service import get_player_snapshot
-from app.services.loyalty_service import get_player_loyalty
+from sync.services.player_service import get_player_snapshot
+from coclib.services.loyalty_service import get_player_loyalty
 
 bp = Blueprint("player", __name__, url_prefix="/player")
