+++ app/extensions.py
@@ -1,9 +1,3 @@
-from flask_apscheduler import APScheduler
-from flask_caching import Cache
-from flask_migrate import Migrate
-from flask_sqlalchemy import SQLAlchemy
+from coclib.extensions import db, cache, scheduler, migrate
 
-db = SQLAlchemy()
-cache = Cache()
-scheduler = APScheduler()
-migrate = Migrate()
+__all__ = ["db", "cache", "scheduler", "migrate"]
