+++ sync/__init__.py
@@ -1,0 +1,1 @@
+# Synchronization service package
